+++ docs/docs/changelog.md
@@ -15,6 +15,8 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 
 ## [Unreleased] - {ISO8601 date}
+### Added
+- Adds lap and session data for FIT files
 
 ## [0.10.0] - 2020-06-01
 ### Added